Strip extra space space after <br> elements
Strip the leading space on the line following a manual line break with
the <br> tag.

In XSL terms, we'd want to ignore the whitespace *around* the <br> tag -
but there is no way to specify this. Hence the slightly more complicated
manual removal if the first preceding sibling of a text node is a <br>
tag.